+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - BREAKING: place the 2 optional args to cursor helpers in an options object
 - BREAKING: removed 2nd argument to upgrade function, which is redundant
 - expose all IDB 3.0 new type additions and normalise existing type usage
+- add optional `options` argument to transaction for IDB 3 transaction options
 - add support for simple managed migrations
 - add signal promises for opened, closed, outdated and blocked
 - add errorevent param to cursor error handler